+++ StudentEnrollment00.InMemory/Events/StudentUpdated.cs
@@ -1,11 +1,10 @@
-namespace StudentEnrollment00.Events;
+namespace StudentEnrollment00.InMemory.Events;
 
-public class StudentCreated : Event
+public class StudentUpdated : Event
 {
     public required Guid StudentId { get; init; }
     public required string FullName { get; init; }
     public required string Email { get; init; }
-    public required DateTime DateOfBirth { get; init; }
 
     public override Guid StreamId => StudentId;
 }
